+++ src/gamelogic/level.py
@@ -8,8 +8,13 @@ from sprites.powerpellet import PowerPellet
 from sprites.ghost import Ghost
 
 
-class Level:  # pylint: disable=too-many-instance-attributes
-    # all instance attributes are necessary for this class
+class Level: # pylint: disable=too-many-instance-attributes
+             # all instance attributes are necessary for this class
+    """Luokka, joka vastaa pelilogiikasta.
+
+    Attributes: level_map: pelattava kenttä ruudukkona.
+                cell_size: luotavan pohjan solujen koko pikseleinä.
+    """
     def __init__(self, level_map, cell_size):
         self.cell_size = cell_size
         self.level = level_map
@@ -19,52 +24,57 @@ class Level:  # pylint: disable=too-many-instance-attributes
         self.power_pellets = pygame.sprite.Group()
         self.ghosts = pygame.sprite.Group()
         self.all_sprites = pygame.sprite.Group()
+        self.pacman = None
         self.score = 0
         self.lives = 3
         self.cleared = 0
         self.timer = None
-        self.initialize_sprites(level_map)
+        self.initialize_sprites()
 
-    def initialize_sprites(self, level_map):  # pylint: disable=too-many-statements
-                                              # statements are required to initialize
-        level_height = len(level_map)
-        level_width = len(level_map[0])
+    def initialize_sprites(self): # pylint: disable=too-many-statements
+                                  # statements are required to initialize
+        """Luodaan spritet käyttäen annettua ruudukkoa.
+
+        Args:
+            level_map: pelattava kenttä ruudukkona.
+        """
+        level_height = len(self.level)
+        level_width = len(self.level[0])
 
         for height in range(level_height):
             for width in range(level_width):
                 normalized_y = height * self.cell_size
                 normalized_x = width * self.cell_size
 
-                if level_map[height][width] == 0:
+                if self.level[height][width] == 0:
                     self.floors.add(Floor(normalized_x, normalized_y))
                     self.pellets.add(Pellet(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 1:
+                elif self.level[height][width] == 1:
                     self.walls.add(Wall(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 2:
+                elif self.level[height][width] == 2:
                     self.floors.add(Floor(normalized_x, normalized_y))
                     self.power_pellets.add(
                         PowerPellet(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 3:
+                elif self.level[height][width] == 3:
                     self.pacman = Pacman(normalized_x, normalized_y)
                     self.floors.add(Floor(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 4:
+                elif self.level[height][width] == 4:
                     self.ghosts.add(Ghost(1, normalized_x, normalized_y))
                     self.floors.add(Floor(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 5:
+                elif self.level[height][width] == 5:
                     self.ghosts.add(Ghost(2, normalized_x, normalized_y))
                     self.floors.add(Floor(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 6:
-                    if self.cleared < 1:
-                        self.ghosts.add(Ghost(3, normalized_x, normalized_y))
+                elif self.level[height][width] == 6:
+                    self.ghosts.add(Ghost(3, normalized_x, normalized_y))
                     self.floors.add(Floor(normalized_x, normalized_y))
 
-                elif level_map[height][width] == 7:
+                elif self.level[height][width] == 7:
                     self.ghosts.add(Ghost(4, normalized_x, normalized_y))
                     self.floors.add(Floor(normalized_x, normalized_y))
 
@@ -78,6 +88,14 @@ class Level:  # pylint: disable=too-many-instance-attributes
         )
 
     def moving_is_possible(self, sprite):
+        """Testaa törmääkö sprite liikkuessaan seinään.
+
+        Args:
+            sprite: mikä tahansa sprite-luokan olio.
+
+        Returns:
+            Boolean arvon törmääkö annettu sprite seinään.
+        """
         sprite.rect.move_ip(
             sprite.direction[0]//sprite.speed, sprite.direction[1]//sprite.speed)
         crashing = pygame.sprite.spritecollide(sprite, self.walls, False)
@@ -87,6 +105,9 @@ class Level:  # pylint: disable=too-many-instance-attributes
         return can_move
 
     def pacman_eats(self):
+        """Tarkistaa pygamen metodilla törmäävätkö
+        pacman-sprite ja pellet- tai powerpellet-sprite ja lisää pisteitä tämän mukaan.
+        """
         if pygame.sprite.spritecollide(self.pacman, self.pellets, True):
             self.score += 10
         elif pygame.sprite.spritecollide(self.pacman, self.power_pellets, True):
@@ -94,16 +115,26 @@ class Level:  # pylint: disable=too-many-instance-attributes
             self.ghosts_are_vulnerable()
 
     def ghosts_are_vulnerable(self):
+        """Asettaa Ghost-luokan spritet vulnerable-tilaan ja vaihtaa kuvan.
+        """
         self.timer = pygame.time.get_ticks()
         for ghost in self.ghosts:
             ghost.vulnerable = True
             ghost.set_image()
 
     def pacman_meets_ghost(self):
+        """Metodi tarkistaa törmäävätkö pacman-sprite ja joku haamuista.
+        Jos haamu on vulnerable-tilassa, se kuolee ja uusi luodaan
+        lähtöpaikkaan self.revive_ghost-metodilla. True palautetaan
+        ainoastaan, kun pacman-sprite menettää elämän.
+
+        Returns:
+            Boolean arvo törmäävätkö pacman- ja ghost-sprite.
+        """
         list_of_colliding = pygame.sprite.spritecollide(
             self.pacman, self.ghosts, False)
         for ghost in list_of_colliding:
-            if ghost.vulnerable == True:
+            if ghost.vulnerable is True:
                 self.score += 100
                 ghost.kill()
                 self.revive_ghost(ghost)
@@ -114,6 +145,12 @@ class Level:  # pylint: disable=too-many-instance-attributes
         return False
 
     def revive_ghost(self, ghost):
+        """Ghost-sprite luodaan uudelleen numeron mukaiseen lähtöruutuun ja speed_up_ghost-
+        metodi tarkistaa tarvitseeko ghost-spriten liikkumista nopeuttaa.
+
+        Args:
+            ghost: Ghost-luokan sprite.
+        """
         starting_point = 150 + 30*ghost.number
         new_ghost = Ghost(ghost.number, starting_point, 210)
         self.speed_up_ghost(new_ghost)
@@ -121,12 +158,20 @@ class Level:  # pylint: disable=too-many-instance-attributes
         self.all_sprites.add(self.ghosts)
 
     def speed_up_ghost(self, ghost):
+        """Jos kenttiä on läpäisty yksi tai enemmän, haamun liikkumista nopeutetaan.
+
+        Args:
+            ghost: Ghost-luokan sprite.
+        """
         if self.cleared >= 1:
-            ghost.speed = ghost.speed-(self.cleared*3)
-            if ghost.speed < 3:
-                ghost.speed = 3
+            ghost.speed = max(ghost.speed-(self.cleared*3), 3)
 
     def move_pacman(self, direction):
+        """Pacman-luokan spriten liikuttaminen.
+
+        Args:
+            direction: Tuple suunnasta, johon spriten tulisi liikkua.
+        """
         self.check_timer()
         self.pacman.new_direction = direction
         if self.centered(self.pacman):
@@ -135,7 +180,8 @@ class Level:  # pylint: disable=too-many-instance-attributes
         if self.moving_is_possible(self.pacman):
             self.pacman.set_image()
             self.pacman.rect.move_ip(
-                self.pacman.direction[0]//self.pacman.speed, self.pacman.direction[1]//self.pacman.speed)
+                self.pacman.direction[0]//self.pacman.speed,
+                self.pacman.direction[1]//self.pacman.speed)
             self.pacman_eats()
 
     def move_ghost(self, sprite):
@@ -163,7 +209,7 @@ class Level:  # pylint: disable=too-many-instance-attributes
         return False
 
     def find_path(self, sprite):
-        if sprite.vulnerable == True:
+        if sprite.vulnerable is True:
             target = [sprite.vulnerable_target[0], sprite.vulnerable_target[1]]
         else:
             target = [self.pacman.rect.x//self.cell_size,
@@ -176,6 +222,15 @@ class Level:  # pylint: disable=too-many-instance-attributes
         return path[0]
 
     def neighbour_is_inside_matrix(self, neighbour, current_cell):
+        """Tarkistetaan kuuluuko annettu solu kentän sisälle.
+
+        Args:
+            neighbour: Naapurisolu.
+            current_cell: Tämän hetkinen solu.
+
+        Returns:
+            Boolean arvo True, jos naapurisolu on ruudukon sisällä, muuten False.
+        """
         if neighbour[0]+current_cell[0] >= 0 and neighbour[0]+current_cell[0] < len(self.level[0]):
             if neighbour[1]+current_cell[1] >= 0 and neighbour[1]+current_cell[1] < len(self.level):
                 return True
@@ -187,7 +242,7 @@ class Level:  # pylint: disable=too-many-instance-attributes
         path = []
         queue.append(starting_cell)
         while queue:  # pylint: disable=too-many-nested-blocks
-            # nested blocks can not be avoided in this algorithm
+                      # nested blocks can not be avoided in this algorithm
             current_cell = queue.popleft()
             visited.append(current_cell)
 
@@ -210,6 +265,9 @@ class Level:  # pylint: disable=too-many-instance-attributes
         return direction
 
     def position_pacman_and_ghosts_to_start(self):
+        """Tuhotaan kaikki Ghost-luokan spritet ja Pacman-luokan sprite ja luodaan ne uudelleen
+        lähtötilanteeseen.
+        """
         for ghost in self.ghosts:
             ghost.kill()
 
@@ -218,13 +276,18 @@ class Level:  # pylint: disable=too-many-instance-attributes
         self.pacman.kill()
         self.all_sprites.remove(self.pacman)
 
-        for x in range(1, 5):
-            self.revive_ghost(Ghost(x, 0, 0))
+        for number in range(1, 5):
+            self.revive_ghost(Ghost(number, 0, 0))
 
         self.pacman = Pacman(8*self.cell_size, 11*self.cell_size)
         self.all_sprites.add(self.pacman)
-    
+
     def all_pellets_eaten(self):
+        """Tarkistetaan onko kaikki Pellet- ja Powerpellet-luokkien spritet on syöty.
+
+        Returns:
+            Boolean arvo True jos kaikki pelletit on syöty, muuten False.
+        """
         if len(self.pellets) == 0 and len(self.power_pellets) == 0:
             return True
         return False